chore: split run tests into modules
